+++ Operatoraccesscontrol/models/CreateOperatorControlDetails.cs
@@ -51,6 +51,10 @@ namespace Oci.OperatoraccesscontrolService.Models
         /// <value>
         /// List of user groups who can approve an access request associated with a resource governed by this operator control.
         /// </value>
+        /// <remarks>
+        /// Required
+        /// </remarks>
+        [Required(ErrorMessage = "ApproverGroupsList is required.")]
         [JsonProperty(PropertyName = "approverGroupsList")]
         public System.Collections.Generic.List<string> ApproverGroupsList { get; set; }
         
@@ -67,8 +71,23 @@ namespace Oci.OperatoraccesscontrolService.Models
         /// will be auto-approved.        
         /// 
         /// </value>
+        /// <remarks>
+        /// Required
+        /// </remarks>
+        [Required(ErrorMessage = "IsFullyPreApproved is required.")]
         [JsonProperty(PropertyName = "isFullyPreApproved")]
         public System.Nullable<bool> IsFullyPreApproved { get; set; }
+        
+        /// <value>
+        /// resourceType for which the OperatorControl is applicable
+        /// </value>
+        /// <remarks>
+        /// Required
+        /// </remarks>
+        [Required(ErrorMessage = "ResourceType is required.")]
+        [JsonProperty(PropertyName = "resourceType")]
+        [JsonConverter(typeof(StringEnumConverter))]
+        public System.Nullable<ResourceTypes> ResourceType { get; set; }
         
         /// <value>
         /// List of emailId.